requirements.txt
pandas>=2.2
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
pyarrow>=15.0
matplotlib>=3.8
seaborn>=0.13
statsmodels>=0.14
scipy>=1.11
geopandas>=1.0
